# perf: reusable cache for scanners