--- a/.github/workflows/CI.yml
+++ b/.github/workflows/CI.yml
@@ -145,6 +145,8 @@ jobs:
         run: pnpm napi create-npm-dirs
       - name: Move artifacts
         run: pnpm artifacts
+      - name: Prepare for publish
+        run: pnpm napi prepublish -t npm --no-gh-release
       - name: List packages
         run: ls -R ./npm
         shell: bash