--- a/package.json
+++ b/package.json
@@ -41,7 +41,10 @@
     "universal": "napi universal",
     "version": "napi version",
     "test": "node --test __test__/index.spec.mjs",
-    "benchmark": "node __test__/benchmark.mjs"
+    "benchmark": "node __test__/benchmark.mjs",
+    "profile": "node --expose-gc __test__/run-profiling.mjs",
+    "profile:baseline": "node --expose-gc __test__/run-profiling.mjs --baseline",
+    "profile:compare": "node --expose-gc __test__/run-profiling.mjs --compare __test__/baseline-performance.json"
   },
   "files": [
     "index.js",