… __test__/baseline-performance.json
… __test__/index.spec.mjs
… __test__/profiling.mjs
… index.d.ts
… src/lib.rs